chore: support current rvv 0.7.1 kernels
